Move spatial functions to separate file
