+++ src/Cypher.ts
@@ -82,15 +82,7 @@ export { plus, minus, divide, multiply, mod, pow } from "./expressions/operation
 // --Functions
 export { CypherFunction as Function } from "./expressions/functions/CypherFunctions";
 
-export {
-    coalesce,
-    point,
-    distance,
-    pointDistance,
-    randomUUID,
-    id,
-    elementId,
-} from "./expressions/functions/CypherFunctions";
+export { coalesce, randomUUID, id, elementId } from "./expressions/functions/CypherFunctions";
 
 export { count, min, max, avg, sum, collect } from "./expressions/functions/AggregationFunctions";
 
@@ -104,7 +96,7 @@ export {
 
 export * from "./expressions/functions/StringFunctions";
 export * from "./expressions/functions/ListFunctions";
-
+export * from "./expressions/functions/SpatialFunctions";
 export * from "./expressions/functions/PathFunctions";
 
 export { any, all, exists, single } from "./expressions/functions/PredicateFunctions";
